EY-3312 Skille ut navnjobb i egen fil
